se agrega feature del 5% de daño extra si esta en terreno enemigo

+++ AlgoChess/src/main/java/edu/fiuba/algoChess/Modelo/salud/Curable.java
@@ -2,6 +2,6 @@ package edu.fiuba.algoChess.Modelo.salud;
 
 public interface Curable {
 
-	public Salud curar(int salud);
+	public Salud curar(double salud);
 
 }
