+++ AlgoChess/src/main/java/edu/fiuba/algoChess/Modelo/salud/Herible.java
@@ -1,5 +1,5 @@
 package edu.fiuba.algoChess.Modelo.salud;
 
 public interface Herible {
-	public Salud herir(int salud);
+	public Salud herir(double salud);
 }
