+++ AlgoChess/src/main/java/edu/fiuba/algoChess/Modelo/salud/SaludLlena.java
@@ -2,17 +2,17 @@ package edu.fiuba.algoChess.Modelo.salud;
 
 public class SaludLlena extends Salud {
 
-	public SaludLlena(int salud){
+	public SaludLlena(double salud){
 		super(salud,salud);
 	}
 
 	@Override
-	public Salud curar(int salud) {
+	public Salud curar(double salud) {
 		return this;
 	}
 
 	@Override
-	public Salud herir(int salud) {
+	public Salud herir(double salud) {
 		setValorActual(this.getValorActual()-salud);
 		return this.getValorActual() == this.getValorCompleto()? this : new SaludMedia(this.getValorCompleto(),this.getValorActual());
 	}
